:recycle: refactor games related features
