+++ src/features/games/components/game-of-the-day/game-of-the-day.component.html
@@ -45,7 +45,7 @@
 
           <!-- BACK -->
           <div class="gotd__card-back">
-            <img class="gotd__flag" src="assets/flags/{{ game.language }}-flag.png" alt="flag">
+            <img class="gotd__flag" src="/assets/flags/{{ game.language }}-flag.png" alt="flag">
             <div class="gotd__info">
               <div class="gotd__info-title">
                 <p class="gotd__name">{{ game.name }}</p>
