+++ src/features/games/components/game-of-the-day/game-of-the-day.component.scss
@@ -1,4 +1,4 @@
-@use "../../core/styles/variables" as variables;
+@use "../../../../core/styles/variables" as variables;
 
 .gotd {
   color: variables.$white;
