+++ src/features/games/components/game-of-the-day/game-of-the-day.component.ts
@@ -8,9 +8,9 @@ import { MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
 
-import { GameCard } from '../models';
-import { CommonFunctionsService } from '../../core/functions/common/common-functions.service';
-import { HttpService } from '../../core/services/http/http.service';
+import { GameCard } from '../../../../components/models';
+import { CommonFunctionsService } from '../../../../core/functions/common/common-functions.service';
+import { HttpService } from '../../../../core/services/http/http.service';
 
 @Component({
   selector: 'app-game-of-the-day',
